+++ src/main/java/ru/skypro/homework/service/impl/UserServiceImpl.java
@@ -28,7 +28,7 @@ public class UserServiceImpl implements UserService {
     private final UserMapper userMapper;
 
     //Находит пользователя по его идентификатору.
-    public UserEntity findUser(Integer userId) {
+    public UserEntity findUser(Long userId) {
         return userRepository.findById(userId)
                 .orElseThrow(() -> new UserNotFoundException("User not found"));
     }
@@ -39,8 +39,8 @@ public class UserServiceImpl implements UserService {
     }
 
     //Находит пользователя по его идентификатору.
-    public UserEntity getUserById(Integer id) {
-        UserEntity byUserId = userRepository.findByUserId(Long.valueOf(id));
+    public UserEntity getUserById(Long id) {
+        UserEntity byUserId = userRepository.findByUserId(id);
         if (byUserId == null) {
             throw new RuntimeException();
         }
@@ -58,8 +58,8 @@ public class UserServiceImpl implements UserService {
     }
 
     //Обновляет информацию о пользователе.
-    public UpdateUserDTO updateUser(Integer userId, UpdateUserDTO updateUser) {
-        UserEntity userEntity = userRepository.findByUserId(Long.valueOf(userId));
+    public UpdateUserDTO updateUser(Long userId, UpdateUserDTO updateUser) {
+        UserEntity userEntity = userRepository.findByUserId(userId);
         if (updateUser.getFirstName() != null) {
             userEntity.setFirstName(updateUser.getFirstName());
         }
@@ -74,17 +74,18 @@ public class UserServiceImpl implements UserService {
     }
 
     //Находит DTO пользователя по его идентификатору.
-    public UserDTO findUserDTO(Integer userId) {
+    public UserDTO findUserDTO(Long userId) {
         return getUserDTO(findUser(userId));
     }
 
     //Обновляет пароль пользователя.
-    public void updatePassword(Integer userId, NewPasswordDTO newPasswordDTO) {
-        UserEntity userEntity = userRepository.findByUserId(Long.valueOf(userId));
+    public void updatePassword(Long userId, NewPasswordDTO newPasswordDTO) {
+        UserEntity userEntity = userRepository.findByUserId(userId);
         if (userEntity.getPassword().equals(newPasswordDTO.getCurrentPassword())) {
             userEntity.setPassword(newPasswordDTO.getNewPassword());
         } else {
             throw new NotEditUserPasswordException("not edited password");
         }
     }
+
 }
